make: Cleaning up VS2017 and 2019 project files
Cleaning up the VS2017 and VS2019 proejct files.
There were a lot of odd differences where only certain build targets were working due to not updating all the preprocessor flags and all the include and linker paths across all builds.
Also fixed a mismatch on static builds in certain cases.
Fixed a major problem where _DEBUG was defined in a release build which broke all release targets. This should never be defined by the prject file as it is autodefined based on the configuration. This caused a lot of headaches to figure out that this was the source of this issue.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>